requirements.txt
# Framework principal da web
Flask==3.0.3
Flask-Cors==4.0.1

# Servidor de produção WSGI (essencial para o Render)
gunicorn==22.0.0

# Cliente oficial da OpenAI
openai==1.30.1

# Cliente oficial do Supabase
supabase==2.5.0

# Para carregar variáveis de ambiente em desenvolvimento local
python-dotenv==1.0.1
